rename non_serializable_kibana_instances to non_serializable_instances, fix test jest test broken by import
